Fixed CRLF skipping in a chunked HTTP request.
The current logic tries to determine wether CRLF-skipping was already done from the size of the body_ variable.
It assumes when body_ is empty, we still have to skip the CRLF. This logic breaks when the first chunk of data just ends
exactly after the CRLF. Such the body_ will remain empty and when the next
chunk arrives, again two bytes are skipped.

One possible solution is, to save explicitly wether CRLF was already skipped or not.
This is done in here.